requirements.txt
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0
torch>=2.0.0
flask==2.0.1
google-cloud-aiplatform==1.25.0 